requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
pydantic[email]==1.10.2
sqlalchemy==1.4.41
sqlalchemy-utils==0.38.3
asyncpg==0.26.0
alembic==1.8.1
python-dotenv==0.21.0
python-multipart==0.0.5
passlib[bcrypt]==1.7.4
aiogram==2.22.1
python-jose[cryptography]==3.3.0
redis==4.3.4
pika==0.5.2
celery==5.2.7